+++ packages/datasource-mongoose/test/integration/future.test.ts
@@ -5,7 +5,7 @@ import { Filter, Projection } from '@forestadmin/datasource-toolkit';
 import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factories__';
 import { Connection, Types } from 'mongoose';
 
-import MongooseDatasource from '../src/datasource';
+import MongooseDatasource from '../../src/datasource';
 import { setupWith2ManyToManyRelations } from './_helpers';
 
 describe('Tests for future versions', () => {
